Revert "Expose public Nutzap profile helper and show details on public profile"
